Fix mouse position floating
When converting from screen to game then coverting back, due to
the scale is not integer(either specify in config file or scale=3
and window size is larger than screen), we can see mouse position
floating when right click to buy potions.

To fix this, when converting coordinates, we need convert the float
point number to nearest integer, instead of simply cut the fraction
part.